refactor: login with discord button
